pyproject.toml
[project]
name = "sightline"
version = "0.1.0"
description = "AI-powered CoD League VOD analysis platform for Faze team strategy"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "opencv-python>=4.9.0",
    "ultralytics>=8.1.0",
    "easyocr>=1.7.0",
    "supervision>=0.19.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "sqlmodel>=0.0.14",
    "openai>=1.10.0",
    "google-genai>=1.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "python-multipart>=0.0.6",
    "yt-dlp>=2024.1.0",
    "umap-learn>=0.5.0",
    "dtw-python>=1.3.0",
    "onnxruntime>=1.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
sightline = "cod_analyst.cli:app"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
